feat: move domain model mappers to specific file
