+++ app/src/main/java/com/github/felipecastilhos/pokedexandroid/features/home/domain/repository/PokemonRepository.kt
@@ -2,11 +2,8 @@ package com.github.felipecastilhos.pokedexandroid.features.home.domain.repositor
 
 import com.github.felipecastilhos.pokedexandroid.GetPokemonQuery
 import com.github.felipecastilhos.pokedexandroid.core.datasource.Resource
-import com.github.felipecastilhos.pokedexandroid.core.logs.LogHandler
 import com.github.felipecastilhos.pokedexandroid.features.home.data.datasource.HomeRemoteDataSource
-import com.github.felipecastilhos.pokedexandroid.features.home.domain.models.FlavorText
 import com.github.felipecastilhos.pokedexandroid.features.home.domain.models.Pokemon
-import com.github.felipecastilhos.pokedexandroid.features.home.domain.models.PokemonTypes
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
@@ -45,31 +42,3 @@ fun Flow<Resource<GetPokemonQuery.GetPokemon?>>.mapToDomainFlow(): Flow<Resource
         }
     }
 }
-
-fun GetPokemonQuery.GetPokemon.mapToDomainModel(): Pokemon = Pokemon(
-    pokedexNumber = this.num,
-    species = this.species,
-    types = this.types.toTypesDomainModel(),
-    height = this.height,
-    weight = this.weight,
-    flavorText = this.flavorTexts.toFlavorListDomainModel()
-)
-
-fun List<GetPokemonQuery.FlavorText>.toFlavorListDomainModel(): List<FlavorText> = map {
-    it.toFlavorListDomainModel()
-}
-
-fun GetPokemonQuery.FlavorText.toFlavorListDomainModel(): FlavorText =
-    FlavorText(game = game, flavor = flavor)
-
-fun List<String>.toTypesDomainModel(): List<PokemonTypes?> = map {
-    it.mapToTypesDomainModel()
-}
-
-fun String.mapToTypesDomainModel(): PokemonTypes? =
-    try {
-        PokemonTypes.valueOf(this)
-    } catch (e: IllegalArgumentException) {
-        LogHandler.d("Não foi converter para modelo de domínio")
-        null
-    }
